🗑[DEL] HomeViewController2 (#6)
