+++ Oliveyoung-iOS/Oliveyoung-iOS/Source/Application/SceneDelegate.swift
@@ -24,7 +24,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
             let window = UIWindow(windowScene: windowScene)
             window.overrideUserInterfaceStyle = UIUserInterfaceStyle.light
 
-            let rootVC = HomeViewController2()
+            let rootVC = HomeViewController()
             let navigationController = UINavigationController(rootViewController: rootVC)
 
             window.rootViewController = rootVC
