+++ Oliveyoung-iOS/Oliveyoung-iOS/Source/Presentation/Search/ViewController/SearchViewController.swift
@@ -124,8 +124,8 @@ extension SearchViewController: UICollectionViewDelegateFlowLayout {
     }
 }
 
-struct ViewControllerPreView:PreviewProvider {
-    static var previews: some View {
-        SearchViewController().toPreview()
-    }
-}
+//struct ViewControllerPreView:PreviewProvider {
+//    static var previews: some View {
+//        SearchViewController().toPreview()
+//    }
+//}
